Use ocamlbuild directly
Solvu-build doesn't support configurable utput directory but we require this is
required for proper integration with Esy.

There's workaround for packages which hardcode output path (or build in source)
but this makes building packages less efficient as Esy copies sources to a
temporary build directory in that case.

Therefore we start to use ocamlbuild directly and while we need to maintain META
and .merlin manually for now this feels cleaner to me for some reason.

I hope in the future we can generate META and .merlin using Esy instead of
relying on a build system.